1) New folder structure + doc
