+++ jsdoc.json
@@ -2,22 +2,30 @@
   "opts": {
     "destination": "./gh-pages/",
     "tutorials": "./docs/",
-    "template": "node_modules/docdash",
-    "private": true
+    "template": "node_modules/docdash"
+  },
+  "plugins": ["plugins/markdown"],
+  "markdown": {
+    "excludeTags": ["author"]
   },
   "source": {
-    "include": [ "./packages/blueconfig/lib" ]
+    "include": [
+      "./packages/blueconfig/lib/",
+      "./packages/blueconfig/lib/format/",
+      "./packages/blueconfig/lib/model/",
+      "./packages/blueconfig/lib/performer/"
+    ]
   },
   "docdash": {
     "sectionOrder": [
-       "Tutorials",
-       "Classes",
-       "Modules",
-       "Externals",
-       "Events",
-       "Namespaces",
-       "Mixins",
-       "Interfaces"
+      "Tutorials",
+      "Classes",
+      "Modules",
+      "Externals",
+      "Events",
+      "Namespaces",
+      "Mixins",
+      "Interfaces"
     ]
   }
 }
