+++ packages/blueconfig-format-with-validator/README.md
@@ -17,8 +17,11 @@ An example `config.js` file:
 ```javascript
 const blueconfig = require('blueconfig');
 
-blueconfig.addFormat(require('blueconfig-format-with-validator').ipaddress);
-blueconfig.addFormat(require('blueconfig-format-with-validator').port);
+blueconfig.addFormats([
+  require('blueconfig-format-with-validator').email,
+  require('blueconfig-format-with-validator').ipaddress,
+  require('blueconfig-format-with-validator').url
+]);
 
 // or :
 // blueconfig.addFormats(require('blueconfig-format-with-validator'));
@@ -26,17 +29,27 @@ blueconfig.addFormat(require('blueconfig-format-with-validator').port);
 // Define a schema
 var config = blueconfig({
   ip: {
-    doc: 'The IP address to bind.',
+    doc: 'The IP address to bind',
     format: 'ipaddress',
     default: '127.0.0.1',
     env: 'IP_ADDRESS',
   },
   port: {
-    doc: 'The port to bind.',
+    doc: 'The port to bind',
     format: 'port',
     default: 8080,
     env: 'PORT',
     arg: 'port'
+  },
+  email: {
+    doc: 'Server mail',
+    format: 'email',
+    default: 'foo@bar.com'
+  },
+  url: {
+    doc: 'Server url',
+    format: 'url',
+    default: 'http://example.com'
   }
 });
 ```
